LayoutTests/inspector/view/synchronous-layout-expected.txt

Testing synchronous View layout operations: updateLayout, updateLayoutIfNeeded.


== Running test suite: View.SynchronousLayout
-- Running test case: View.updateLayout
Update layout for attached view.
PASS: View should do an initial layout.
PASS: View should update its layout once.
PASS: View should not have a pending layout.
Update layout for detached view.
PASS: View should do an initial layout.
PASS: View should update its layout once.
PASS: View should not have a pending layout.

-- Running test case: View.updateLayout.propogateToSubviews
Update parent view layout.
PASS: Child view should do an initial layout.
PASS: Child view should update its layout once.

-- Running test case: View.updateLayoutIfNeeded
PASS: View should update if an initial layout never happened.
PASS: View should update if a layout is pending.
PASS: View should not update if no layout is pending.
